requirements.txt
cycler==0.10.0
decorator==4.1.2
dlib==19.6.1
functools32==3.2.3.post2
matplotlib==2.0.2
networkx==1.11
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.1
six==1.10.0
subprocess32==3.2.7
